fix: clean exports, colocate render options to templates
