+++ app/routes/case-studies.tsx
@@ -1,43 +1,11 @@
-import {
-  BLOCKS,
-  INLINES,
-  type Block,
-  type Inline,
-} from "@contentful/rich-text-types"
 import { json, type LoaderFunctionArgs } from "@remix-run/node"
 import { type MetaFunction, useLoaderData } from "@remix-run/react"
-import { type ReactNode } from "react"
 import { type RootLoader } from "~/root"
 import { invariantResponse } from "~/utils/invariant.server"
 import { getSocialMetas } from "~/utils/seo"
 import { getCaseStudiesPage } from "~/models/case-studies.server"
 import { CaseStudiesPage } from "~/ui/templates/CaseStudiesPage"
 import { getEPARegionsWithCaseStudies } from "~/models/epa-region.server"
-
-export const richTextRenderOptions = {
-  renderNode: {
-    [INLINES.HYPERLINK]: (node: Block | Inline, children: ReactNode) => {
-      const { data } = node
-      const { uri } = data
-      return (
-        <a
-          className="text-primary underline dark:text-gray-400"
-          target="_blank"
-          rel="noreferrer"
-          href={uri}
-        >
-          {children}
-        </a>
-      )
-    },
-    [BLOCKS.PARAGRAPH]: (node: Block | Inline, children: ReactNode) => {
-      return <p className="mb-4 text-base leading-relaxed">{children}</p>
-    },
-    [BLOCKS.HEADING_2]: (node: Block | Inline, children: ReactNode) => {
-      return <h2 className="mb-5 text-3xl">{children}</h2>
-    },
-  },
-}
 
 export const loader = async (_: LoaderFunctionArgs) => {
   const caseStudiesPage = await getCaseStudiesPage("4d7g9VYoTaIuUCuVEJu4U2")
