+++ app/routes/team.$memberSlug.tsx
@@ -1,16 +1,9 @@
 import {
-  BLOCKS,
-  INLINES,
-  type Block,
-  type Inline,
-} from "@contentful/rich-text-types"
-import {
   json,
   type LoaderFunctionArgs,
   type MetaFunction,
 } from "@remix-run/node"
 import { useLoaderData } from "@remix-run/react"
-import { type ReactNode } from "react"
 import {
   getTeamMemberBySlug,
   getTeamMembers,
@@ -25,65 +18,6 @@ import type { SEOHandle } from "@nasa-gcn/remix-seo"
 import { Show500 } from "~/ui/templates/500"
 import { z } from "zod"
 import { serverOnly$ } from "vite-env-only"
-
-export const richTextRenderOptions = {
-  renderNode: {
-    [INLINES.HYPERLINK]: (node: Block | Inline, children: ReactNode) => {
-      const { data } = node
-      const { uri } = data
-      return (
-        <a
-          className="text-primary underline dark:text-gray-400"
-          target="_blank"
-          rel="noreferrer"
-          href={uri}
-        >
-          {children}
-        </a>
-      )
-    },
-    [BLOCKS.PARAGRAPH]: (node: Block | Inline, children: ReactNode) => {
-      return (
-        <p className="mb-4 text-base leading-relaxed text-black">{children}</p>
-      )
-    },
-    [BLOCKS.HEADING_2]: (node: Block | Inline, children: ReactNode) => {
-      return (
-        <h2 className="mb-5 text-3xl font-bold dark:text-gray-200">
-          {children}
-        </h2>
-      )
-    },
-    [BLOCKS.HEADING_3]: (node: Block | Inline, children: ReactNode) => {
-      return (
-        <h3 className="mb-4 text-2xl font-bold dark:text-gray-200">
-          {children}
-        </h3>
-      )
-    },
-    [BLOCKS.HEADING_4]: (node: Block | Inline, children: ReactNode) => {
-      return (
-        <h4 className="mb-4 text-xl uppercase dark:text-gray-200">
-          {children}
-        </h4>
-      )
-    },
-    [BLOCKS.HEADING_5]: (node: Block | Inline, children: ReactNode) => {
-      return (
-        <h5 className="mb-4 text-lg font-bold dark:text-gray-200">
-          {children}
-        </h5>
-      )
-    },
-    [BLOCKS.HEADING_6]: (node: Block | Inline, children: ReactNode) => {
-      return (
-        <h6 className="text-md mb-4 font-bold uppercase dark:text-gray-200">
-          {children}
-        </h6>
-      )
-    },
-  },
-}
 
 export const loader = async ({ params }: LoaderFunctionArgs) => {
   const { memberSlug } = params
